@@ -14,34 +14,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.5",
-    "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^4.3.5",
-    "@angular/compiler": "^4.3.5",
-    "@angular/core": "^4.3.5",
-    "@angular/forms": "^4.3.5",
-    "@angular/http": "^4.3.5",
-    "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.3.5",
-    "@angular/platform-browser-dynamic": "^4.3.5",
-    "@angular/router": "^4.3.5",
-    "angularfire2": "^4.0.0-rc.2",
+    "@angular/animations": "^5.0.0",
+    "@angular/cdk": "^5.0.0-rc0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/material": "^5.0.0-rc0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/platform-server": "^5.0.0",
+    "@angular/router": "^5.0.0",
+    "angularfire2": "^5.0.0-rc.3",
     "core-js": "^2.5.0",
     "firebase": "^4.3.0",
     "hammerjs": "^2.0.8",
     "immutable": "^3.8.1",
     "lodash": "^4.17.4",
-    "rxjs": "^5.4.3",
+    "rxjs": "^5.5.2",
     "typed-immutable-record": "0.0.6",
     "zone.js": "^0.8.16"
   },
   "devDependencies": {
     "@angular/cli": "^1.3.1",
-    "@angular/compiler-cli": "^4.3.5",
+    "@angular/compiler-cli": "^5.0.0",
     "@types/jasmine": "2.5.38",
     "@types/lodash": "^4.14.57",
     "@types/node": "~6.0.60",
-    "codelyzer": "^3.1.2",
+    "codelyzer": "^4.0.1",
     "jasmine-core": "~2.5.2",
     "jasmine-spec-reporter": "~3.2.0",
     "karma": "~1.4.1",
@@ -53,6 +54,6 @@
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "^5.6.0",
-    "typescript": "^2.4.2"
+    "typescript": "2.4.2"
   }
 }
