@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {ReplaySubject} from 'rxjs/ReplaySubject';
 import {Observable} from 'rxjs/Observable';
-import {AngularFireDatabase, FirebaseObjectObservable} from 'angularfire2/database';
+import {AngularFireDatabase} from 'angularfire2/database';
 import {Category} from '../model/category';
 import {Catalog} from '../model/catalog';
 import {Item} from '../model/item';
@@ -15,10 +15,10 @@ import {CategoryRecord} from '../model/category-record';
 @Injectable()
 export class CatalogService {
   private catalogObservable: ReplaySubject<Catalog> = new ReplaySubject();
-  private item: FirebaseObjectObservable<any[]>;
+  private item: Observable<any[]>;
 
   constructor(ad: AngularFireDatabase) {
-    this.item = ad.object('/catalog');
+    this.item = ad.object('/catalog').valueChanges();
     this.item.subscribe(snapshot => this.buildCatalog(snapshot));
   }
 
